"check_style_score -> 0 of 5 passed. style score: 0/1 "

+++ cspp2-assignments/m6/Assignment-2/Solution.java
@@ -22,13 +22,13 @@ final class Solution {
    *
    * @return     Matrix of the rounded elements
    */
-  static int[][] roundHundred(final int[][] a, final int rows, final int columns) {
+  static double[][] roundHundred(final int[][] a, final int rows, final int columns) {
 
     // write your code here
     double[][] b = new double[rows][columns];
     for (int i = 0; i < rows; i++) {
       for (int j = 0; j < columns; j++) {
-        b[i][j] = a[i][j]/100;
+        b[i][j] = a[i][j] / 100;
       }
     }
     return b;
@@ -48,7 +48,7 @@ final class Solution {
         a[i][j] = scan.nextInt();
       }
     }
-    int[][] b = roundHundred(a, m, n);
+    double[][] b = roundHundred(a, m, n);
     for (int i = 0; i < m; i++) {
       for (int j = 0; j < n - 1; j++) {
         System.out.print(b[i][j] + " ");
